功能完善

+++ internal/controller/corp/corp_v1_delete.go
@@ -2,12 +2,18 @@ package corp
 
 import (
 	"context"
+	"github.com/gogf/gf/v2/frame/g"
 	"github.com/iimeta/fastapi-admin/internal/service"
 
 	"github.com/iimeta/fastapi-admin/api/corp/v1"
 )
 
 func (c *ControllerV1) Delete(ctx context.Context, req *v1.DeleteReq) (res *v1.DeleteRes, err error) {
+
+	if service.Session().IsUserRole(ctx) {
+		g.RequestFromCtx(ctx).Response.WriteJson(g.Map{"code": 401, "message": "Unauthorized"})
+		return
+	}
 
 	err = service.Corp().Delete(ctx, req.Id)
